+++ pages/index.tsx
@@ -1,65 +1,26 @@
 import type { NextPage } from "next";
-import { Center, Container, Stack } from "@mantine/core";
-import useSWR, { useSWRConfig } from "swr";
-import AddItem from "../components/addItem";
-import Item from "../components/item";
-import { Todo } from "../lib/db";
+import Link from "next/link";
+import { Center, Container, Title, Button, Space, Text } from "@mantine/core";
+import { AiFillGithub } from "react-icons/ai";
+import useSWR from "swr";
 import SEO from "../components/seo";
 
-interface PostProps {
-  todos: Todo[];
-}
-
-const Home: NextPage<PostProps> = ({ todos }) => {
-  const { mutate } = useSWRConfig();
-  const { data, error } = useSWR<Todo[]>(`/api/todo`, (url: string) =>
-    fetch(url).then((res) => res.json())
-  );
-
-  const handleAdd = async (description: string) => {
-    if (!description) return;
-    await fetch("/api/todo", {
-      method: "POST",
-      body: JSON.stringify({ description }),
-    });
-    mutate(`/api/todo`, data);
-  };
-
-  const handleDelete = async (id: number) => {
-    if (!id) return;
-    await fetch("/api/todo", {
-      method: "DELETE",
-      body: JSON.stringify({ id }),
-    });
-    mutate(`/api/todo`, data);
-  };
-
-  const handleSave = async (id: number, description: string, done: boolean) => {
-    if (!id) return;
-    if (!description) return;
-    await fetch("/api/todo", {
-      method: "PATCH",
-      body: JSON.stringify({ id, description, done }),
-    });
-    mutate(`/api/todo`, data);
-  };
-
+const Home: NextPage<any> = () => {
   return (
     <>
       <SEO title="Teste SEO" description="teste Description 66" />
       <Container>
-        <Center sx={{ marginTop: "14px" }}>
-          <Stack spacing="lg" sx={{ width: "100%", maxWidth: "500px" }}>
-            {data?.map((item, index) => (
-              <Item
-                {...item}
-                delete={handleDelete}
-                save={handleSave}
-                key={index}
-              />
-            ))}
-            <AddItem create={handleAdd} />
-          </Stack>
+        <Center
+          sx={{ marginTop: "14px", display: "flex", flexDirection: "column" }}
+        >
+          <Space h="lg" />
+          <Text> Conecte-se e comece a terminar suas tarefas</Text>
+          <Space h="lg" />
+          <Link href="/dashboard" passHref>
+            <Button leftIcon={<AiFillGithub size={18} />} component="a" compact>
+              Entrar com Github
+            </Button>
+          </Link>
         </Center>
       </Container>
     </>
